separated account creation, reworked test structure
