+++ src/tests/Feature/FinancialOperations/OperationsOverviewFilteringTest.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Tests\Feature\FinancialAccount;
+namespace Tests\Feature\FinancialOperations;
 
-use App\Http\Controllers\FinancialOperations\OperationsOverviewController;
 use App\Models\Account;
 use App\Models\FinancialOperation;
 use App\Models\User;
@@ -15,7 +14,7 @@ use Tests\Util\HiddenMembersAccessor;
 /**
  * These tests must be run on a seeded database, as they generate plenty of models with foreign keys.
  */
-class OperationsFilteringTest extends TestCase
+class OperationsOverviewFilteringTest extends TestCase
 {
     use DatabaseTransactions;
